@@ -1,5 +1,5 @@
 {
-  "name": "car-car-app",
+  "name": "my-i18n-app",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,38 +9,31 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@radix-ui/react-accordion": "^1.2.11",
-    "@radix-ui/react-avatar": "^1.1.10",
-    "@radix-ui/react-checkbox": "^1.3.2",
-    "@radix-ui/react-dialog": "^1.1.14",
-    "@radix-ui/react-hover-card": "^1.1.14",
-    "@radix-ui/react-popover": "^1.1.14",
     "@radix-ui/react-select": "^2.2.5",
-    "@radix-ui/react-slider": "^1.3.5",
     "@radix-ui/react-slot": "^1.2.3",
-    "@supabase/auth-helpers-nextjs": "^0.10.0",
-    "@tanstack/react-query": "^5.81.5",
+    "@tanstack/react-query": "^5.83.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cmdk": "^1.1.1",
-    "dayjs": "^1.11.13",
+    "i18next": "^25.3.2",
     "lucide-react": "^0.525.0",
-    "next": "15.3.4",
+    "next": "15.4.2",
     "next-themes": "^0.4.6",
     "pg": "^8.16.3",
-    "radix-ui": "^1.4.2",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-i18next": "^15.6.0",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
-    "@types/node": "^20",
+    "@types/node": "^20.19.9",
     "@types/pg": "^8.15.4",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@types/react-i18next": "^7.8.3",
     "tailwindcss": "^4",
-    "tw-animate-css": "^1.3.4",
+    "ts-node": "^10.9.2",
+    "tw-animate-css": "^1.3.5",
     "typescript": "^5"
   }
 }
