@@ -1,8 +1,9 @@
-import type {Metadata} from "next";
-import {Nunito} from "next/font/google";
+import type { Metadata } from "next";
+import { Nunito } from 'next/font/google';
 import "./globals.css";
-import {ThemeProvider} from "next-themes";
 import Providers from "@/app/Providers";
+import {ThemeProvider} from "next-themes";
+import Header from "@/components/shared/Header/Header";
 
 const nunito = Nunito({
     variable: '--font-nunito',
@@ -11,26 +12,27 @@ const nunito = Nunito({
 });
 
 export const metadata: Metadata = {
-    title: "CarCar Home",
-    description: "Home page",
+  title: "Create Next App",
+  description: "Generated by create next app",
 };
 
 export default function RootLayout({
-                                       children,
-                                   }: Readonly<{
-    children: React.ReactNode;
+  children,
+}: Readonly<{
+  children: React.ReactNode;
 }>) {
-    return (
-        <html lang="en">
-        <body
-            className={`${nunito.className} antialiased`}
-        >
-        <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
-            <Providers>
-                {children}
-            </Providers>
-        </ThemeProvider>
-        </body>
-        </html>
-    );
+  return (
+    <html lang="en">
+      <body
+        className={`${nunito.className} antialiased`}
+      >
+      <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+          <Providers>
+              <Header />
+              {children}
+          </Providers>
+      </ThemeProvider>
+      </body>
+    </html>
+  );
 }
